Clarity fix
Renamed SyncServer to make it less confusing
