+++ src/main/java/com/superzanti/serversync/util/PathUtils.java
@@ -10,7 +10,11 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.stream.Stream;
-
+/**
+ * Helper for working with paths and directories
+ * @author Rheimus
+ *
+ */
 public class PathUtils {
 
 	private static StringBuilder pathBuilder;
